+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Article paragraph numbering is no longer continuous
 ### Changed
 - Use `AuModalContainer` component instead of  #ember-appuniversum-wormhole element in dummy app
+- Improved documentation of the plugins
 
 ## [3.0.0] - 2023-02-27