+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Feature: make citation use the new link node
 - BREAKING: citation mark has been removed
+- Table of contents now is able to be inserted in the corresponding place instead of always being inserted at position 0
 - Change paragraph symbol to §
 - Update schema on dummy page to make articles insertable in empty document
 - Add padding to structure card