+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change paragraph symbol to §
 - Update schema on dummy page to make articles insertable in empty document
 - Add padding to structure card
+- Placeholder text when inserting date
 
 ## [3.1.0] - 2023-03-02