+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change paragraph symbol to §
 - Update schema on dummy page to make articles insertable in empty document
 - Add padding to structure card
+- Placeholder text when inserting date
 
 ### Removed:
 - Removal of prosemirror-plugin dependency of `CitationPlugin::CitationInsert` component.